% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Summary Statistics Columns}
\usage{
summary_statistics(data, x, y)
}
\arguments{
\item{data}{is the dataset you wish to create the calculated columns within. data should contain the columns of interest}

\item{x}{is the column you wish to calculate the summary statistics on}

\item{y}{is the column that you will be grouping your data with}
}
\value{
5 columns of summary statistics on x data, with one row for each correlating to the y grouping within the dataset data.
}
\description{
This function will create columns in the dataset calculating the mean, median, standard
deviation, max and min of a variable column based on being grouped by a second column
}
\examples{
letter <- data.frame(type = c("A","A","A", "O", "O","O","P","P","P"),number =c(1,7,4,3,45,7,9,0,2))
summary_statistics(letter, number, type)
}
